Everything up to and including a <script> tag was omitted from RSS
The "[\s\S]*" in the regular expression was swallowing up the entire
article up to the <script> tag. Removing this allows the regex to remove
just the tag itself. (I also changed the greedy "*" quantifier to the
non-greedy "*?"; this will ensure that only one <script> at a time is
removed. Otherwise, if there were more than one script tag in an article
then everything between the first and last <script> would have been cut
out.)